+++ packages/assets-controllers/src/TokenRatesController.test.ts
@@ -1,51 +1,20 @@
 import { NetworksTicker, toHex } from '@metamask/controller-utils';
+import type { Hex } from '@metamask/utils';
 import nock from 'nock';
 import { useFakeTimers } from 'sinon';
 
 import { advanceTime } from '../../../tests/helpers';
+import type {
+  AbstractTokenPricesService,
+  TokenPrice,
+  TokenPricesByTokenContractAddress,
+} from './token-prices-service/abstract-token-prices-service';
 import { TokenRatesController } from './TokenRatesController';
 
-const COINGECKO_API = 'https://api.coingecko.com/api/v3';
-const COINGECKO_ETH_PATH = '/simple/token_price/ethereum';
-const COINGECKO_MATIC_PATH = '/simple/token_price/polygon-pos-network';
-const COINGECKO_ASSETS_PATH = '/asset_platforms';
-const COINGECKO_SUPPORTED_CURRENCIES = '/simple/supported_vs_currencies';
 const ADDRESS = '0x01';
-
 const defaultSelectedAddress = '0x0000000000000000000000000000000000000001';
 
 describe('TokenRatesController', () => {
-  beforeEach(() => {
-    nock(COINGECKO_API)
-      .get(COINGECKO_SUPPORTED_CURRENCIES)
-      .reply(200, ['eth', 'usd', 'dai'])
-      .get(COINGECKO_ASSETS_PATH)
-      .reply(200, [
-        {
-          id: 'binance-smart-chain',
-          chain_identifier: 56,
-          name: 'Binance Smart Chain',
-          shortname: 'BSC',
-        },
-        {
-          id: 'ethereum',
-          chain_identifier: 1,
-          name: 'Ethereum',
-          shortname: '',
-        },
-        {
-          id: 'polygon-pos-network',
-          chain_identifier: 137,
-          name: 'Polygon',
-          shortname: 'MATIC',
-        },
-      ]);
-
-    nock('https://min-api.cryptocompare.com')
-      .get('/data/price?fsym=ETH&tsyms=ETH')
-      .reply(200, { ETH: 1 });
-  });
-
   afterEach(() => {
     jest.restoreAllMocks();
   });
@@ -70,6 +39,7 @@ describe('TokenRatesController', () => {
         onPreferencesStateChange: jest.fn(),
         onTokensStateChange: jest.fn(),
         onNetworkStateChange: jest.fn(),
+        tokenPricesService: buildMockTokenPricesService(),
       });
       expect(controller.state).toStrictEqual({
         contractExchangeRates: {},
@@ -86,6 +56,7 @@ describe('TokenRatesController', () => {
         onPreferencesStateChange: jest.fn(),
         onTokensStateChange: jest.fn(),
         onNetworkStateChange: jest.fn(),
+        tokenPricesService: buildMockTokenPricesService(),
       });
       expect(controller.config).toStrictEqual({
         interval: 180000,
@@ -110,6 +81,7 @@ describe('TokenRatesController', () => {
         onPreferencesStateChange: jest.fn(),
         onTokensStateChange: jest.fn(),
         onNetworkStateChange: jest.fn(),
+        tokenPricesService: buildMockTokenPricesService(),
       });
 
       await advanceTime({ clock, duration: 500 });
@@ -145,6 +117,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange: jest.fn(),
             onTokensStateChange,
             onNetworkStateChange: jest.fn(),
+            tokenPricesService: buildMockTokenPricesService(),
           },
           {
             allTokens: {
@@ -201,6 +174,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange: jest.fn(),
             onTokensStateChange,
             onNetworkStateChange: jest.fn(),
+            tokenPricesService: buildMockTokenPricesService(),
           },
           {
             allDetectedTokens,
@@ -242,6 +216,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange: jest.fn(),
             onTokensStateChange,
             onNetworkStateChange: jest.fn(),
+            tokenPricesService: buildMockTokenPricesService(),
           },
           {
             allTokens: {
@@ -287,6 +262,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange: jest.fn(),
             onTokensStateChange,
             onNetworkStateChange: jest.fn(),
+            tokenPricesService: buildMockTokenPricesService(),
           },
           {
             allDetectedTokens: {
@@ -348,6 +324,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         await controller.start();
         const updateExchangeRatesSpy = jest
@@ -378,6 +355,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         await controller.start();
         const updateExchangeRatesSpy = jest
@@ -408,6 +386,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         await controller.start();
         jest.spyOn(controller, 'updateExchangeRates').mockResolvedValue();
@@ -436,6 +415,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         await controller.start();
         jest.spyOn(controller, 'updateExchangeRates').mockResolvedValue();
@@ -464,6 +444,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         await controller.start();
         const updateExchangeRatesSpy = jest
@@ -496,6 +477,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         const updateExchangeRatesSpy = jest
           .spyOn(controller, 'updateExchangeRates')
@@ -525,6 +507,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         const updateExchangeRatesSpy = jest
           .spyOn(controller, 'updateExchangeRates')
@@ -554,6 +537,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         jest.spyOn(controller, 'updateExchangeRates').mockResolvedValue();
 
@@ -581,6 +565,7 @@ describe('TokenRatesController', () => {
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange,
+          tokenPricesService: buildMockTokenPricesService(),
         });
         jest.spyOn(controller, 'updateExchangeRates').mockResolvedValue();
 
@@ -625,6 +610,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange,
             onTokensStateChange: jest.fn(),
             onNetworkStateChange: jest.fn(),
+            tokenPricesService: buildMockTokenPricesService(),
           },
           {
             allTokens: {
@@ -667,6 +653,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange,
             onTokensStateChange: jest.fn(),
             onNetworkStateChange: jest.fn(),
+            tokenPricesService: buildMockTokenPricesService(),
           },
           {
             allTokens: {
@@ -714,6 +701,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange,
             onTokensStateChange: jest.fn(),
             onNetworkStateChange: jest.fn(),
+            tokenPricesService: buildMockTokenPricesService(),
           },
           {
             allTokens: {
@@ -753,12 +741,9 @@ describe('TokenRatesController', () => {
 
     describe('start', () => {
       it('should poll and update rate in the right interval', async () => {
-        const fetchSpy = jest
-          .spyOn(globalThis, 'fetch')
-          .mockImplementation(() => {
-            throw new Error('Network error');
-          });
         const interval = 100;
+        const tokenPricesService = buildMockTokenPricesService();
+        jest.spyOn(tokenPricesService, 'fetchTokenPrices');
         const controller = new TokenRatesController(
           {
             interval,
@@ -769,6 +754,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange: jest.fn(),
             onTokensStateChange: jest.fn(),
             onNetworkStateChange: jest.fn(),
+            tokenPricesService,
           },
           {
             allTokens: {
@@ -782,24 +768,21 @@ describe('TokenRatesController', () => {
         );
 
         await controller.start();
-        expect(fetchSpy).toHaveBeenCalledTimes(1);
+        expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(1);
 
         await advanceTime({ clock, duration: interval });
-        expect(fetchSpy).toHaveBeenCalledTimes(2);
+        expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(2);
 
         await advanceTime({ clock, duration: interval });
-        expect(fetchSpy).toHaveBeenCalledTimes(3);
+        expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(3);
       });
     });
 
     describe('stop', () => {
       it('should stop polling', async () => {
-        const fetchSpy = jest
-          .spyOn(globalThis, 'fetch')
-          .mockImplementation(() => {
-            throw new Error('Network error');
-          });
         const interval = 100;
+        const tokenPricesService = buildMockTokenPricesService();
+        jest.spyOn(tokenPricesService, 'fetchTokenPrices');
         const controller = new TokenRatesController(
           {
             interval,
@@ -810,6 +793,7 @@ describe('TokenRatesController', () => {
             onPreferencesStateChange: jest.fn(),
             onTokensStateChange: jest.fn(),
             onNetworkStateChange: jest.fn(),
+            tokenPricesService,
           },
           {
             allTokens: {
@@ -823,12 +807,12 @@ describe('TokenRatesController', () => {
         );
 
         await controller.start();
-        expect(fetchSpy).toHaveBeenCalledTimes(1);
+        expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(1);
 
         controller.stop();
 
         await advanceTime({ clock, duration: interval });
-        expect(fetchSpy).toHaveBeenCalledTimes(1);
+        expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(1);
       });
     });
   });
@@ -846,6 +830,8 @@ describe('TokenRatesController', () => {
 
     it('should poll on the right interval', async () => {
       const interval = 100;
+      const tokenPricesService = buildMockTokenPricesService();
+      jest.spyOn(tokenPricesService, 'fetchTokenPrices');
       const controller = new TokenRatesController({
         interval,
         chainId: '0x2',
@@ -860,26 +846,27 @@ describe('TokenRatesController', () => {
             ticker: NetworksTicker.mainnet,
           },
         }),
+        tokenPricesService,
       });
-      const updateExchangeRatesByChainIdSpy = jest
-        .spyOn(controller, 'updateExchangeRatesByChainId')
-        .mockResolvedValue();
 
       controller.startPollingByNetworkClientId('mainnet', {
         tokenAddresses: ['0x0'],
       });
       await advanceTime({ clock, duration: 0 });
-      expect(updateExchangeRatesByChainIdSpy).toHaveBeenCalledTimes(1);
+      expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(1);
 
       await advanceTime({ clock, duration: interval });
-      expect(updateExchangeRatesByChainIdSpy).toHaveBeenCalledTimes(2);
+      expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(2);
 
       await advanceTime({ clock, duration: interval });
-      expect(updateExchangeRatesByChainIdSpy).toHaveBeenCalledTimes(3);
+      expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(3);
     });
 
     it('should update state on poll', async () => {
       const interval = 100;
+      const tokenPricesService = buildMockTokenPricesService({
+        fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+      });
       const controller = new TokenRatesController({
         interval,
         chainId: '0x2',
@@ -894,11 +881,7 @@ describe('TokenRatesController', () => {
             ticker: NetworksTicker.mainnet,
           },
         }),
-      });
-      jest.spyOn(controller, 'getChainSlug').mockResolvedValue('ethereum');
-      jest.spyOn(controller, 'fetchAndMapExchangeRates').mockResolvedValue({
-        '0x02': 0.001,
-        '0x03': 0.002,
+        tokenPricesService,
       });
 
       controller.startPollingByNetworkClientId('mainnet', {
@@ -918,6 +901,8 @@ describe('TokenRatesController', () => {
 
     it('should stop polling', async () => {
       const interval = 100;
+      const tokenPricesService = buildMockTokenPricesService();
+      jest.spyOn(tokenPricesService, 'fetchTokenPrices');
       const controller = new TokenRatesController({
         interval,
         chainId: '0x2',
@@ -932,43 +917,45 @@ describe('TokenRatesController', () => {
             ticker: NetworksTicker.mainnet,
           },
         }),
+        tokenPricesService,
       });
-      const updateExchangeRatesByChainIdSpy = jest
-        .spyOn(controller, 'updateExchangeRatesByChainId')
-        .mockResolvedValue();
 
       const pollingToken = controller.startPollingByNetworkClientId('mainnet', {
         tokenAddresses: ['0x0'],
       });
       await advanceTime({ clock, duration: 0 });
-      expect(updateExchangeRatesByChainIdSpy).toHaveBeenCalledTimes(1);
+      expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(1);
 
       controller.stopPollingByPollingToken(pollingToken);
 
       await advanceTime({ clock, duration: interval });
-      expect(updateExchangeRatesByChainIdSpy).toHaveBeenCalledTimes(1);
+      expect(tokenPricesService.fetchTokenPrices).toHaveBeenCalledTimes(1);
     });
   });
 
   describe('updateExchangeRates', () => {
     it('should not update exchange rates if legacy polling is disabled', async () => {
-      const controller = new TokenRatesController({
-        chainId: '0x1',
-        ticker: NetworksTicker.mainnet,
-        selectedAddress: defaultSelectedAddress,
-        onPreferencesStateChange: jest.fn(),
-        onTokensStateChange: jest.fn(),
-        onNetworkStateChange: jest.fn(),
-        getNetworkClientById: jest.fn(),
-      });
-      controller.disabled = true;
-
-      const updateExchangeRatesByChainIdSpy = jest
-        .spyOn(controller, 'updateExchangeRatesByChainId')
-        .mockResolvedValue();
+      const tokenPricesService = buildMockTokenPricesService();
+      jest.spyOn(tokenPricesService, 'fetchTokenPrices');
+      const controller = new TokenRatesController(
+        {
+          chainId: '0x1',
+          ticker: NetworksTicker.mainnet,
+          selectedAddress: defaultSelectedAddress,
+          onPreferencesStateChange: jest.fn(),
+          onTokensStateChange: jest.fn(),
+          onNetworkStateChange: jest.fn(),
+          getNetworkClientById: jest.fn(),
+          tokenPricesService,
+        },
+        {
+          disabled: true,
+        },
+      );
 
       await controller.updateExchangeRates();
-      expect(updateExchangeRatesByChainIdSpy).not.toHaveBeenCalled();
+
+      expect(tokenPricesService.fetchTokenPrices).not.toHaveBeenCalled();
     });
 
     it('should update legacy state after updateExchangeRatesByChainId', async () => {
@@ -981,6 +968,7 @@ describe('TokenRatesController', () => {
           onTokensStateChange: jest.fn(),
           onNetworkStateChange: jest.fn(),
           getNetworkClientById: jest.fn(),
+          tokenPricesService: buildMockTokenPricesService(),
         },
         {
           allTokens: {
@@ -1018,7 +1006,7 @@ describe('TokenRatesController', () => {
       expect(updateExchangeRatesByChainIdSpy).toHaveBeenCalledWith({
         chainId: '0x1',
         nativeCurrency: NetworksTicker.mainnet,
-        tokenAddresses: ['0x123', ADDRESS],
+        tokenContractAddresses: ['0x123', ADDRESS],
       });
 
       expect(controller.state.contractExchangeRates).toStrictEqual({
@@ -1029,7 +1017,7 @@ describe('TokenRatesController', () => {
   });
 
   describe('updateExchangeRatesByChainId', () => {
-    it('should not update state if no tokenAddresses are provided', async () => {
+    it('should not update state if no token contract addresses are provided', async () => {
       const controller = new TokenRatesController({
         interval: 100,
         chainId: '0x2',
@@ -1039,26 +1027,96 @@ describe('TokenRatesController', () => {
         onTokensStateChange: jest.fn(),
         onNetworkStateChange: jest.fn(),
         getNetworkClientById: jest.fn(),
+        tokenPricesService: buildMockTokenPricesService(),
       });
 
       expect(controller.state.contractExchangeRates).toStrictEqual({});
       await controller.updateExchangeRatesByChainId({
         chainId: '0x1',
         nativeCurrency: 'ETH',
-        tokenAddresses: [],
+        tokenContractAddresses: [],
       });
       expect(controller.state.contractExchangeRatesByChainId).toStrictEqual({});
     });
 
-    it('should update all rates', async () => {
-      const tokenAddress = '0x89d24A6b4CcB1B6fAA2625fE562bDD9a23260359';
-      nock(COINGECKO_API)
-        .get(
-          `${COINGECKO_ETH_PATH}?contract_addresses=${tokenAddress}&vs_currencies=eth`,
-        )
-        .reply(200, {
-          '0x89d24a6b4ccb1b6faa2625fe562bdd9a23260359': { eth: 0.00561045 },
-        });
+    it('should not update state when disabled', async () => {
+      const tokenContractAddress = '0x89d24A6b4CcB1B6fAA2625fE562bDD9a23260359';
+      const controller = new TokenRatesController(
+        {
+          interval: 100,
+          chainId: '0x2',
+          ticker: 'ticker',
+          selectedAddress: '0xdeadbeef',
+          onPreferencesStateChange: jest.fn(),
+          onTokensStateChange: jest.fn(),
+          onNetworkStateChange: jest.fn(),
+          getNetworkClientById: jest.fn(),
+          tokenPricesService: buildMockTokenPricesService(),
+        },
+        { disabled: true },
+      );
+      expect(controller.state.contractExchangeRatesByChainId).toStrictEqual({});
+
+      await controller.updateExchangeRatesByChainId({
+        chainId: '0x1',
+        nativeCurrency: 'ETH',
+        tokenContractAddresses: [tokenContractAddress],
+      });
+
+      expect(controller.state.contractExchangeRatesByChainId).toStrictEqual({});
+    });
+
+    it('should update exchange rates for the given token addresses to undefined when the given chain ID is not supported by the Price API', async () => {
+      const controller = new TokenRatesController(
+        {
+          chainId: '0x2',
+          ticker: 'ticker',
+          selectedAddress: '0xdeadbeef',
+          onPreferencesStateChange: jest.fn(),
+          onTokensStateChange: jest.fn(),
+          onNetworkStateChange: jest.fn(),
+          getNetworkClientById: jest.fn(),
+          tokenPricesService: buildMockTokenPricesService({
+            validateChainIdSupported(chainId: unknown): chainId is Hex {
+              return chainId !== '0x9999999999';
+            },
+          }),
+        },
+        {},
+        {
+          contractExchangeRatesByChainId: {
+            '0x9999999999': {
+              MATIC: {
+                '0x02': 0.01,
+                '0x03': 0.02,
+                '0x04': 0.03,
+              },
+            },
+          },
+        },
+      );
+
+      await controller.updateExchangeRatesByChainId({
+        chainId: '0x9999999999',
+        nativeCurrency: 'MATIC',
+        tokenContractAddresses: ['0x02', '0x03'],
+      });
+
+      expect(controller.state.contractExchangeRatesByChainId).toStrictEqual({
+        '0x9999999999': {
+          MATIC: {
+            '0x02': undefined,
+            '0x03': undefined,
+            '0x04': 0.03,
+          },
+        },
+      });
+    });
+
+    it('should update exchange rates when native currency is supported by the Price API', async () => {
+      const tokenPricesService = buildMockTokenPricesService({
+        fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+      });
       const controller = new TokenRatesController({
         interval: 100,
         chainId: '0x2',
@@ -1068,66 +1126,34 @@ describe('TokenRatesController', () => {
         onTokensStateChange: jest.fn(),
         onNetworkStateChange: jest.fn(),
         getNetworkClientById: jest.fn(),
+        tokenPricesService,
       });
 
       expect(controller.state.contractExchangeRates).toStrictEqual({});
       await controller.updateExchangeRatesByChainId({
         chainId: '0x1',
         nativeCurrency: 'ETH',
-        tokenAddresses: [tokenAddress],
+        tokenContractAddresses: ['0xAAA'],
       });
-      expect(
-        Object.keys(controller.state.contractExchangeRatesByChainId['0x1'].ETH),
-      ).toContain(tokenAddress);
-      expect(
-        controller.state.contractExchangeRatesByChainId['0x1'].ETH[
-          tokenAddress
-        ],
-      ).toBeGreaterThan(0);
+      expect(controller.state.contractExchangeRatesByChainId).toStrictEqual({
+        '0x1': {
+          ETH: {
+            '0xAAA': 0.001,
+          },
+        },
+      });
     });
 
-    it('should handle balance not found in API', async () => {
-      const controller = new TokenRatesController({
-        chainId: '0x2',
-        ticker: 'ticker',
-        selectedAddress: '0xdeadbeef',
-        onPreferencesStateChange: jest.fn(),
-        onTokensStateChange: jest.fn(),
-        onNetworkStateChange: jest.fn(),
-        getNetworkClientById: jest.fn(),
-      });
-      expect(controller.state.contractExchangeRates).toStrictEqual({});
-      jest.spyOn(controller, 'fetchExchangeRate').mockRejectedValue({
-        error: 'Not Found',
-        message: 'Not Found',
-      });
-
-      const result = controller.updateExchangeRatesByChainId({
-        chainId: '0x1',
-        nativeCurrency: 'ETH',
-        tokenAddresses: ['0x0'],
-      });
-
-      await expect(result).rejects.not.toThrow();
-    });
-
-    it('should update exchange rates when native currency is not supported by coingecko', async () => {
-      nock(COINGECKO_API)
-        .get(`${COINGECKO_MATIC_PATH}`)
-        .query({ contract_addresses: '0x02,0x03', vs_currencies: 'eth' })
-        .reply(200, {
-          '0x02': {
-            eth: 0.001, // token value in terms of ETH
-          },
-          '0x03': {
-            eth: 0.002,
-          },
-        });
-
+    it('should update exchange rates when native currency is not supported by the Price API', async () => {
       nock('https://min-api.cryptocompare.com')
-        .get('/data/price?fsym=ETH&tsyms=MATIC')
-        .reply(200, { MATIC: 0.5 }); // .5 eth to 1 matic
-
+        .get('/data/price?fsym=ETH&tsyms=LOL')
+        .reply(200, { LOL: 0.5 });
+      const tokenPricesService = buildMockTokenPricesService({
+        fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+        validateCurrencySupported(currency: unknown): currency is string {
+          return currency !== 'LOL';
+        },
+      });
       const controller = new TokenRatesController({
         chainId: '0x2',
         ticker: 'ticker',
@@ -1136,104 +1162,36 @@ describe('TokenRatesController', () => {
         onTokensStateChange: jest.fn(),
         onNetworkStateChange: jest.fn(),
         getNetworkClientById: jest.fn(),
+        tokenPricesService,
       });
 
       await controller.updateExchangeRatesByChainId({
-        chainId: toHex(137),
-        nativeCurrency: 'MATIC',
-        tokenAddresses: ['0x02', '0x03'],
+        chainId: '0x1',
+        nativeCurrency: 'LOL',
+        tokenContractAddresses: ['0x02', '0x03'],
       });
 
       expect(controller.state.contractExchangeRatesByChainId).toStrictEqual({
-        [toHex(137)]: {
-          MATIC: {
-            '0x02': 0.0005, // token value in terms of matic = (token value in eth) * (eth value in matic) = .001 * .5
+        '0x1': {
+          LOL: {
+            // token price in LOL = (token price in ETH) * (ETH value in LOL)
+            '0x02': 0.0005,
             '0x03': 0.001,
           },
         },
       });
     });
-
-    it('should update exchange rates with undefined when chain is not supported by coingecko', async () => {
-      const controller = new TokenRatesController({
-        chainId: '0x2',
-        ticker: 'ticker',
-        selectedAddress: '0xdeadbeef',
-        onPreferencesStateChange: jest.fn(),
-        onTokensStateChange: jest.fn(),
-        onNetworkStateChange: jest.fn(),
-        getNetworkClientById: jest.fn(),
-      });
-      jest.spyOn(controller, 'getChainSlug').mockResolvedValue('');
-
-      await controller.updateExchangeRatesByChainId({
-        chainId: toHex(137),
-        nativeCurrency: 'MATIC',
-        tokenAddresses: ['0x02', '0x03'],
-      });
-
-      expect(controller.state.contractExchangeRatesByChainId).toStrictEqual({
-        [toHex(137)]: {
-          MATIC: {
-            '0x02': undefined,
-            '0x03': undefined,
-          },
-        },
-      });
-    });
-  });
-
-  describe('getChainSlug', () => {
-    it('returns the chain slug for the chain id', async () => {
-      const controller = new TokenRatesController({
-        chainId: '0x2',
-        ticker: 'ticker',
-        selectedAddress: '0xdeadbeef',
-        onPreferencesStateChange: jest.fn(),
-        onTokensStateChange: jest.fn(),
-        onNetworkStateChange: jest.fn(),
-        getNetworkClientById: jest.fn(),
-      });
-
-      expect(await controller.getChainSlug('0x1')).toBe('ethereum');
-      expect(await controller.getChainSlug(toHex(56))).toBe(
-        'binance-smart-chain',
-      );
-      expect(await controller.getChainSlug(toHex(137))).toBe(
-        'polygon-pos-network',
-      );
-    });
-
-    it('returns null if there is no chain slug for the chain id', async () => {
-      const controller = new TokenRatesController({
-        chainId: '0x2',
-        ticker: 'ticker',
-        selectedAddress: '0xdeadbeef',
-        onPreferencesStateChange: jest.fn(),
-        onTokensStateChange: jest.fn(),
-        onNetworkStateChange: jest.fn(),
-        getNetworkClientById: jest.fn(),
-      });
-
-      expect(await controller.getChainSlug('0x2')).toBeNull();
-    });
   });
 
   describe('fetchAndMapExchangeRates', () => {
-    describe('native currency is supported', () => {
+    describe('when the native currency is supported', () => {
       it('returns the exchange rates directly', async () => {
-        nock(COINGECKO_API)
-          .get(`${COINGECKO_ETH_PATH}`)
-          .query({ contract_addresses: '0x02,0x03', vs_currencies: 'eth' })
-          .reply(200, {
-            '0x02': {
-              eth: 0.001,
-            },
-            '0x03': {
-              eth: 0.002,
-            },
-          });
-
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+          validateCurrencySupported(currency: unknown): currency is string {
+            return currency === 'ETH';
+          },
+        });
         const controller = new TokenRatesController({
           chainId: '0x2',
           ticker: 'ticker',
@@ -1242,12 +1200,15 @@ describe('TokenRatesController', () => {
           onTokensStateChange: jest.fn(),
           onNetworkStateChange: jest.fn(),
           getNetworkClientById: jest.fn(),
+          tokenPricesService,
         });
 
         const contractExchangeRates = await controller.fetchAndMapExchangeRates(
-          'ETH',
-          'ethereum',
-          ['0x02', '0x03'],
+          {
+            nativeCurrency: 'ETH',
+            chainId: '0x1',
+            tokenContractAddresses: ['0x02', '0x03'],
+          },
         );
 
         expect(contractExchangeRates).toStrictEqual({
@@ -1257,24 +1218,17 @@ describe('TokenRatesController', () => {
       });
     });
 
-    describe('native currency is not supported', () => {
+    describe('when the native currency is not supported', () => {
       it('returns the exchange rates using ETH as a fallback currency', async () => {
-        nock(COINGECKO_API)
-          .get(`${COINGECKO_ETH_PATH}`)
-          .query({ contract_addresses: '0x02,0x03', vs_currencies: 'eth' })
-          .reply(200, {
-            '0x02': {
-              eth: 0.001,
-            },
-            '0x03': {
-              eth: 0.002,
-            },
-          });
-
         nock('https://min-api.cryptocompare.com')
           .get('/data/price?fsym=ETH&tsyms=LOL')
           .reply(200, { LOL: 0.5 });
-
+        const tokenPricesService = buildMockTokenPricesService({
+          fetchTokenPrices: fetchTokenPricesWithIncreasingPriceForEachToken,
+          validateCurrencySupported(currency: unknown): currency is string {
+            return currency !== 'LOL';
+          },
+        });
         const controller = new TokenRatesController({
           chainId: '0x2',
           ticker: 'ticker',
@@ -1283,51 +1237,49 @@ describe('TokenRatesController', () => {
           onTokensStateChange: jest.fn(),
           onNetworkStateChange: jest.fn(),
           getNetworkClientById: jest.fn(),
+          tokenPricesService,
         });
 
         const contractExchangeRates = await controller.fetchAndMapExchangeRates(
-          'LOL',
-          'ethereum',
-          ['0x02', '0x03'],
+          {
+            chainId: '0x1',
+            nativeCurrency: 'LOL',
+            tokenContractAddresses: ['0x02', '0x03'],
+          },
         );
 
         expect(contractExchangeRates).toStrictEqual({
+          // token price in LOL = (token price in ETH) * (ETH value in LOL)
           '0x02': 0.0005,
           '0x03': 0.001,
         });
       });
 
       it('returns the an empty object when market does not exist for pair', async () => {
-        nock(COINGECKO_API)
-          .get(`${COINGECKO_ETH_PATH}`)
-          .query({ contract_addresses: '0x02,0x03', vs_currencies: 'eth' })
-          .reply(200, {
-            '0x02': {
-              eth: 0.001,
-            },
-            '0x03': {
-              eth: 0.002,
-            },
-          });
+        nock('https://min-api.cryptocompare.com')
+          .get('/data/price?fsym=ETH&tsyms=LOL')
+          .replyWithError(
+            new Error('market does not exist for this coin pair'),
+          );
 
+        const tokenPricesService = buildMockTokenPricesService();
         const controller = new TokenRatesController({
           chainId: '0x2',
-          ticker: 'ticker',
+          ticker: 'ETH',
           selectedAddress: '0xdeadbeef',
           onPreferencesStateChange: jest.fn(),
           onTokensStateChange: jest.fn(),
           onNetworkStateChange: jest.fn(),
           getNetworkClientById: jest.fn(),
+          tokenPricesService,
         });
-        jest
-          .spyOn(controller, 'fetchExchangeRate')
-          .mockRejectedValue(
-            new Error('market does not exist for this coin pair'),
-          );
+
         const contractExchangeRates = await controller.fetchAndMapExchangeRates(
-          'LOL',
-          'ethereum',
-          ['0x02', '0x03'],
+          {
+            chainId: '0x1',
+            nativeCurrency: 'LOL',
+            tokenContractAddresses: ['0x02', '0x03'],
+          },
         );
 
         expect(contractExchangeRates).toStrictEqual({});
@@ -1335,3 +1287,61 @@ describe('TokenRatesController', () => {
     });
   });
 });
+
+/**
+ * Builds a mock token prices service.
+ *
+ * @param overrides - The properties of the token prices service you want to
+ * provide explicitly.
+ * @returns The built mock token prices service.
+ */
+function buildMockTokenPricesService(
+  overrides: Partial<AbstractTokenPricesService> = {},
+): AbstractTokenPricesService {
+  return {
+    async fetchTokenPrices() {
+      return {};
+    },
+    validateChainIdSupported(_chainId: unknown): _chainId is Hex {
+      return true;
+    },
+    validateCurrencySupported(_currency: unknown): _currency is string {
+      return true;
+    },
+    ...overrides,
+  };
+}
+
+/**
+ * A version of the token prices service `fetchTokenPrices` method where the
+ * price of each given token is incremented by one.
+ *
+ * @param args - The arguments to this function.
+ * @param args.tokenContractAddresses - The token contract addresses.
+ * @param args.currency - The currency.
+ * @returns The token prices.
+ */
+async function fetchTokenPricesWithIncreasingPriceForEachToken<
+  TokenAddress extends Hex,
+  Currency extends string,
+>({
+  tokenContractAddresses,
+  currency,
+}: {
+  tokenContractAddresses: TokenAddress[];
+  currency: Currency;
+}) {
+  return tokenContractAddresses.reduce<
+    Partial<TokenPricesByTokenContractAddress<TokenAddress, Currency>>
+  >((obj, tokenContractAddress, i) => {
+    const tokenPrice: TokenPrice<TokenAddress, Currency> = {
+      tokenContractAddress,
+      value: (i + 1) / 1000,
+      currency,
+    };
+    return {
+      ...obj,
+      [tokenContractAddress]: tokenPrice,
+    };
+  }, {}) as TokenPricesByTokenContractAddress<TokenAddress, Currency>;
+}