+++ packages/selected-network-controller/src/SelectedNetworkController.ts
@@ -166,11 +166,10 @@ export class SelectedNetworkController extends BaseController<
     }
 
     this.update((state) => {
-      if (state.perDomainNetwork) {
-        state.domains[domain] = networkClientId;
-        return;
+      state.domains[domain] = networkClientId;
+      if (!state.perDomainNetwork) {
+        state.domains[METAMASK_DOMAIN] = networkClientId;
       }
-      state.domains[METAMASK_DOMAIN] = networkClientId;
     });
   }
 
@@ -204,5 +203,19 @@ export class SelectedNetworkController extends BaseController<
     }
 
     return networkProxy;
+  }
+
+  setPerDomainNetwork(enabled: boolean) {
+    this.update((state) => {
+      state.perDomainNetwork = enabled;
+      return state;
+    });
+    Object.keys(this.state.domains).forEach((domain) => {
+      // when perDomainNetwork is false, getNetworkClientIdForDomain always returns the networkClientId for the domain 'metamask'
+      this.setNetworkClientIdForDomain(
+        domain,
+        this.getNetworkClientIdForDomain(domain),
+      );
+    });
   }
 }